autoposter/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autoposter"
version = "0.1.0"
description = "Semi-automated TLA+ monthly development update blog post generator"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "openai>=1.30",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pygal>=3.0",
    "python-dateutil>=2.9",
]

[project.scripts]
devupdate = "autoposter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
